+++ src/com/apmath/loans/application/v1/models/incoming/Info.kt
@@ -1,0 +1,3 @@
+package com.apmath.loans.application.v1.models.incoming
+
+data class Info(val version: String)
